separate out ThrusterEngine
