+++ src/main/java/net/jcm/vsch/blocks/entity/ThrusterBlockEntity.java
@@ -1,6 +1,8 @@
 package net.jcm.vsch.blocks.entity;
 
 import net.jcm.vsch.blocks.entity.template.AbstractThrusterBlockEntity;
+import net.jcm.vsch.blocks.thruster.ThrusterEngine;
+import net.jcm.vsch.blocks.thruster.ThrusterEngineContext;
 import net.jcm.vsch.config.VSCHConfig;
 
 import net.minecraft.core.BlockPos;
